修改了日志布局 加入了警报窗口中显示收到的图片的功能，待测试

+++ app/src/main/java/com/SmartHome/SmartHomeDemo/fragments/HomeFragment/HomeFurnitureSpecific.java
@@ -794,7 +794,7 @@
                     case "air_conditioner":
                         TextView arg2TextView = this.getView().findViewById(R.id.arg2_item);
                         if (arg2TextView != null) {
-                            arg2TextView.setText(furnitureItem.getAcTemp() + "℃");
+                            arg2TextView.setText((int)furnitureItem.getAcTemp() + "℃");
                         }
                         break;
                     case "light":
